+++ src/subcommands/runtime/aliasResolution.ts
@@ -1,8 +1,8 @@
 import { RuntimeEngineInfo } from "../../../../lms-shared-types/dist/types/RuntimeEngine.js";
 import { compareVersions } from "../../compareVersions.js";
+import { UserInputError } from "../../types/UserInputError.js";
 import { AliasField, BuiltAlias, fallbackAlias } from "./aliasGeneration.js";
 import { AliasGroup } from "./aliasGrouping.js";
-import { UserInputError } from "./UserInputError.js";
 
 // Returns list of all matching engines
 export function resolveAlias(
