+++ src/subcommands/runtime/list.ts
@@ -11,9 +11,9 @@ import {
 import { compareVersions } from "../../compareVersions.js";
 import { addCreateClientOptions, createClient } from "../../createClient.js";
 import { addLogLevelOptions, createLogger } from "../../logLevel.js";
+import { UserInputError } from "../../types/UserInputError.js";
 import { fallbackAlias } from "./aliasGeneration.js";
 import { AliasGroup } from "./aliasGrouping.js";
-import { UserInputError } from "./UserInputError.js";
 
 export interface RuntimeEngineDisplayInfo {
   specifier: RuntimeEngineSpecifier;
